Refactored dot notation in file names.
