+++ app/components/server-view/server-list/serverList.js
@@ -6,8 +6,8 @@ import { Redirect } from 'react-router'
 import VirtualList from '../../controls/VirtualizedList/VirtualList'
 import ServerListItem from './serverListItem'
 
-import treeViewSpanTypes from '../../../constants/treeViewSpanTypes'
-import treeViewItemTypes from '../../../constants/treeViewItemTypes'
+import treeViewSpanTypes from '../../../constants/treeViewSpanType'
+import treeViewItemTypes from '../../../constants/treeViewItemType'
 import appConstants from '../../../constants/appConstants'
 import defaultServerConfig from '../../../constants/defaultServerConfig'
 import serverTreeViewNodeType from '../../../constants/serverTreeViewNodeType'
