+++ app/reducers/keyView.js
