+++ app/reducers/mainLayout.js
