+++ app/reducers/serverSettings.js
